fix deprecation compiler warnings
