+++ adverse-events/src/main/java/gov/nih/ncats/adverseevents/AdverseEventsApplication.java
@@ -3,11 +3,7 @@ package gov.nih.ncats.adverseevents;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 import gov.hhs.gsrs.adverseevents.AdverseEventDataSourceConfig;
 import gov.hhs.gsrs.adverseevents.adverseeventcvm.EnableAdverseEventCvm;
@@ -35,7 +31,7 @@ import gsrs.EnableGsrsScheduler;
 @SpringBootApplication
 @EntityScan(basePackages ={"ix","gsrs", "gov.nih.ncats", "gov.hhs.gsrs"} )
 @EnableGsrsApi(indexValueMakerDetector = EnableGsrsApi.IndexValueMakerDetector.CONF,
-		additionalDatabaseSourceConfigs= {AdverseEventDataSourceConfig.class}
+                additionalDatabaseSourceConfigs= {AdverseEventDataSourceConfig.class}
 )
 @EnableGsrsJpaEntities
 @EnableGsrsLegacyAuthentication
@@ -47,20 +43,7 @@ import gsrs.EnableGsrsScheduler;
 
 public class AdverseEventsApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(AdverseEventsApplication.class, args);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurerAdapter() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**")
-						.allowedOrigins("*")
-						.allowedMethods( "POST","GET", "OPTIONS", "DELETE", "PUT")
-				;
-			}
-		};
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(AdverseEventsApplication.class, args);
+    }
 }
