+++ adverse-events/src/main/java/gov/nih/ncats/adverseevents/config/MvcConfiguration.java
@@ -1,0 +1,17 @@
+package gov.nih.ncats.adverseevents.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class MvcConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("*")
+                .allowedMethods( "POST","GET", "OPTIONS", "DELETE", "PUT")
+                ;
+    }
+}
