+++ applications/src/main/java/gov/nih/ncats/application/ApplicationApplication.java
@@ -3,11 +3,7 @@ package gov.nih.ncats.application;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 import gov.hhs.gsrs.application.ApplicationDataSourceConfig;
 import gov.hhs.gsrs.application.application.EnableApplication;
@@ -40,7 +36,7 @@ import gsrs.EnableGsrsScheduler;
 @SpringBootApplication
 @EntityScan(basePackages ={"ix","gsrs", "gov.nih.ncats", "gov.hhs.gsrs"} )
 @EnableGsrsApi(indexValueMakerDetector = EnableGsrsApi.IndexValueMakerDetector.CONF,
-		additionalDatabaseSourceConfigs= {ApplicationDataSourceConfig.class}
+                additionalDatabaseSourceConfigs= {ApplicationDataSourceConfig.class}
 )
 @EnableGsrsJpaEntities
 @EnableGsrsLegacyAuthentication
@@ -52,23 +48,8 @@ import gsrs.EnableGsrsScheduler;
 
 public class ApplicationApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ApplicationApplication.class, args);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurerAdapter() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**")
-                      //  .allowedOrigins("http://localhost:4200")
-                        .allowedMethods("GET", "PUT", "POST", "PATCH", "DELETE", "OPTIONS")
-                        .allowedHeaders("origin", "Content-Type", "Authorization", "Accept", "Accept-Language", "X-Authorization", "X-Requested-With", "auth-username", "auth-password", "auth-token", "auth-key", "auth-token")
-                        .allowCredentials(false).maxAge(300);
-
-			}
-		};
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ApplicationApplication.class, args);
+    }
 }
