+++ clinical-trials/src/main/java/gov/hhs/gsrs/clinicaltrial/ClinicalTrialApplication.java
@@ -12,10 +12,6 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 @EnableClinicalTrialUS
 @EnableClinicalTrialEurope
@@ -24,9 +20,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 // explicit reference to FDA-specific systems.
 
 @SpringBootApplication
-@EnableGsrsApi(
-		indexValueMakerDetector = EnableGsrsApi.IndexValueMakerDetector.CONF,
-		additionalDatabaseSourceConfigs = {ClinicalTrialDataSourceConfig.class}
+@EnableGsrsApi(indexValueMakerDetector = EnableGsrsApi.IndexValueMakerDetector.CONF,
+                additionalDatabaseSourceConfigs = {ClinicalTrialDataSourceConfig.class}
 )
 @EnableGsrsJpaEntities
 @EnableGsrsLegacyAuthentication
@@ -52,24 +47,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 // @EnableAutoConfiguration(exclude = {  DataSourceAutoConfiguration.class })
 public class ClinicalTrialApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ClinicalTrialApplication.class, args);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurerAdapter() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-			 registry.addMapping("/**")
-                        // .allowedOrigins("http://localhost:4200")
-			 .allowedOrigins("*")
-                        .allowedMethods("GET", "PUT", "POST", "PATCH", "DELETE", "OPTIONS")
-                        .allowedHeaders("origin", "Content-Type", "Authorization", "Accept", "Accept-Language", "X-Authorization", "X-Requested-With", "auth-username", "auth-password", "auth-token", "auth-key", "auth-token")
-                        .allowCredentials(false).maxAge(300);
-
-			}
-		};
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ClinicalTrialApplication.class, args);
+    }
 }
