+++ frontend/src/main/java/gov/nih/ncats/gsrsfrontend/GsrsFrontendApplication.java
@@ -2,10 +2,6 @@ package gov.nih.ncats.gsrsfrontend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 @SpringBootApplication
 public class GsrsFrontendApplication {
@@ -13,7 +9,4 @@ public class GsrsFrontendApplication {
     public static void main(String[] args) {
         SpringApplication.run(GsrsFrontendApplication.class, args);
     }
-
-
-
 }
