+++ impurities/src/main/java/gov/nih/ncats/impurities/ImpuritiesApplication.java
@@ -9,10 +9,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 //TODO: This may be an issue, and we may need to refactor to make less
 // explicit reference to FDA-specific systems.
@@ -25,7 +21,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 @SpringBootApplication
 @EntityScan(basePackages ={"ix","gsrs", "gov.nih.ncats", "gov.hhs.gsrs"} )
 @EnableGsrsApi(indexValueMakerDetector = EnableGsrsApi.IndexValueMakerDetector.CONF,
-		additionalDatabaseSourceConfigs= {ImpuritiesDataSourceConfig.class}
+                additionalDatabaseSourceConfigs= {ImpuritiesDataSourceConfig.class}
 )
 @EnableGsrsJpaEntities
 @EnableGsrsLegacyAuthentication
@@ -37,23 +33,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 
 public class ImpuritiesApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ImpuritiesApplication.class, args);
-	}
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurerAdapter() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/**")
-                      //  .allowedOrigins("http://localhost:4200")
-                        .allowedMethods("GET", "PUT", "POST", "PATCH", "DELETE", "OPTIONS")
-                        .allowedHeaders("origin", "Content-Type", "Authorization", "Accept", "Accept-Language", "X-Authorization", "X-Requested-With", "auth-username", "auth-password", "auth-token", "auth-key", "auth-token")
-                        .allowCredentials(false).maxAge(300);
-
-			}
-		};
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ImpuritiesApplication.class, args);
+    }
 }
